move test results types to common lib
