+++ packages/esm-patient-common-lib/src/index.ts
@@ -3,5 +3,5 @@ export * from './empty-state';
 export * from './error-state';
 export * from './createDashboardLink';
 export * from './openWorkspaceTab';
-export * from './types';
+export * from './types/index';
 export * from './launchStartVisitPrompt';
