+++ packages/esm-patient-test-results-app/src/desktopView/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Switch, Route, BrowserRouter, useParams } from 'react-router-dom';
+import { Switch, Route, BrowserRouter } from 'react-router-dom';
 import { testResultsBasePath } from '../helpers';
 import DesktopView from './desktopView';
