+++ packages/esm-patient-test-results-app/src/overview/helpers.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './lab-results.scss';
 import { TableRow } from 'carbon-components-react';
-import { OBSERVATION_INTERPRETATION } from '../loadPatientTestData/helpers';
+import { OBSERVATION_INTERPRETATION } from '@openmrs/esm-patient-common-lib';
 import Information16 from '@carbon/icons-react/es/information/16';
 import dayjs from 'dayjs';
